requirements.txt
streamlit
qrcode[pil]
opencv-python-headless
Pillow
supabase